services/worker/pyproject.toml
[project]
name = "pulse-worker"
version = "0.1.0"
description = "Pulse worker service"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.7",
  "google-genai>=1.12.1",
  "httpx>=0.28.1",
  "inngest>=0.4.18",
  "pydantic-ai>=0.0.24",
  "pydantic-settings>=2.7.1",
  "selectolax>=0.3.27",
  "sentence-transformers>=3.3.1",
  "uvicorn>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.3",
  "ruff>=0.8.6"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

